+++ ProjectManagement/ProjectManagement.Client/Helper/ExcelHelper.cs
@@ -62,8 +62,8 @@ namespace ProjectManagement.Client.Helper
                     else
                     {
                         task.Metadata.QuantityParam = ConstValues.FixedQ;
-                        _ = double.TryParse(quantityStr, out double quantity);
-                        _ = double.TryParse(priceStr, out double price);
+                        _ = decimal.TryParse(quantityStr, out decimal quantity);
+                        _ = decimal.TryParse(priceStr, out decimal price);
                         task.Metadata.Quantity = quantity;
                         task.Metadata.PriceSubDB = price;
                     }
